+++ src/FileRenamer/RuntimeException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * General Exception
+ * RuntimeException for FileRenamer
  * 
  * @link        https://github.com/picamator/FileRenamer
  * @license     http://opensource.org/licenses/BSD-3-Clause New BSD License
@@ -8,7 +8,7 @@
 
 namespace FileRenamer;
 
-class GeneralException extends \Exception
+class RuntimeException extends \RuntimeException
 {
     
 }