+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 CHANGELOG
 =========
+v1.3.1 (2014-11-09)
+-------------------
++ Updated unitTest configuration
++ Used RuntimeExeption instead of Exception
++ Renamed tests, docs and configs folders to test, doc and config
+
 v1.3.0 (2014-01-06)
 -------------------
 * Updated Autoload
@@ -11,18 +17,15 @@ v1.2.0 (2013-09-29)
 + UnitTest
 * Set file name changelog, license and readme to upper case
 
-
 v1.1.0 (2013-07-10)
 -------------------
 + Strategies: Translit German, Portuguese to Latin
 + Strategy Hash
 * Changed folder structure for Translit strategies
 
-
 v1.0.1 (2013-07-05)
 -------------------
 * Bug fix: in autoload algorithm when root folder is not named as FileRenamer
-
 
 v1.0.0 (2013-07-04)
 -------------------